pyproject.toml
[tool.poetry]
name = "book-shop-microservice"
version = "0.1.0"
description = ""
authors = ["AleksandrPrishepov <sanyaprishepov89@gmail.com>"]
readme = "README.md"
packages = [{include = "book_shop_microservice"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.2"}
sqlalchemy = "^2.0.15"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


